Move the ConnectionPool to util and restructure a bunch of stuff so we can make util public.
The ConnPool will probably move somewhere else in the longer term.
