+++ src/util.rs
@@ -1,13 +1,9 @@
-use std::ops::{Bound, RangeBounds};
-
-use bao_tree::{io::round_up_to_chunks, ChunkNum, ChunkRanges};
-use range_collections::{range_set::RangeSetEntry, RangeSet2};
-
-pub mod channel;
-pub(crate) mod connection_pool;
+//! Utilities
+pub(crate) mod channel;
+pub mod connection_pool;
 pub(crate) mod temp_tag;
 
-pub mod serde {
+pub(crate) mod serde {
     // Module that handles io::Error serialization/deserialization
     pub mod io_error_serde {
         use std::{fmt, io};
@@ -218,74 +214,7 @@ pub mod serde {
     }
 }
 
-pub trait ChunkRangesExt {
-    fn last_chunk() -> Self;
-    fn chunk(offset: u64) -> Self;
-    fn bytes(ranges: impl RangeBounds<u64>) -> Self;
-    fn chunks(ranges: impl RangeBounds<u64>) -> Self;
-    fn offset(offset: u64) -> Self;
-}
-
-impl ChunkRangesExt for ChunkRanges {
-    fn last_chunk() -> Self {
-        ChunkRanges::from(ChunkNum(u64::MAX)..)
-    }
-
-    /// Create a chunk range that contains a single chunk.
-    fn chunk(offset: u64) -> Self {
-        ChunkRanges::from(ChunkNum(offset)..ChunkNum(offset + 1))
-    }
-
-    /// Create a range of chunks that contains the given byte ranges.
-    /// The byte ranges are rounded up to the nearest chunk size.
-    fn bytes(ranges: impl RangeBounds<u64>) -> Self {
-        round_up_to_chunks(&bounds_from_range(ranges, |v| v))
-    }
-
-    /// Create a range of chunks from u64 chunk bounds.
-    ///
-    /// This is equivalent but more convenient than using the ChunkNum newtype.
-    fn chunks(ranges: impl RangeBounds<u64>) -> Self {
-        bounds_from_range(ranges, ChunkNum)
-    }
-
-    /// Create a chunk range that contains a single byte offset.
-    fn offset(offset: u64) -> Self {
-        Self::bytes(offset..offset + 1)
-    }
-}
-
-// todo: move to range_collections
-pub(crate) fn bounds_from_range<R, T, F>(range: R, f: F) -> RangeSet2<T>
-where
-    R: RangeBounds<u64>,
-    T: RangeSetEntry,
-    F: Fn(u64) -> T,
-{
-    let from = match range.start_bound() {
-        Bound::Included(start) => Some(*start),
-        Bound::Excluded(start) => {
-            let Some(start) = start.checked_add(1) else {
-                return RangeSet2::empty();
-            };
-            Some(start)
-        }
-        Bound::Unbounded => None,
-    };
-    let to = match range.end_bound() {
-        Bound::Included(end) => end.checked_add(1),
-        Bound::Excluded(end) => Some(*end),
-        Bound::Unbounded => None,
-    };
-    match (from, to) {
-        (Some(from), Some(to)) => RangeSet2::from(f(from)..f(to)),
-        (Some(from), None) => RangeSet2::from(f(from)..),
-        (None, Some(to)) => RangeSet2::from(..f(to)),
-        (None, None) => RangeSet2::all(),
-    }
-}
-
-pub mod outboard_with_progress {
+pub(crate) mod outboard_with_progress {
     use std::io::{self, BufReader, Read};
 
     use bao_tree::{
@@ -433,7 +362,7 @@ pub mod outboard_with_progress {
     }
 }
 
-pub mod sink {
+pub(crate) mod sink {
     use std::{future::Future, io};
 
     use irpc::RpcMessage;
